==> Bank/00 - Log.ipynb <==
## Data Validation

- Size of both datasets
    - Application: 647
    - Loan: 1285

- Empty Fields
    - Only the secondary phone type
    
- Emails
    - No email correspond to the users name - inputed to the test - no use

- Address
    - The CEP, Address and Lat, Long match - tried 5 different address
    - used google maps.
    
## Structure
- pre process
    - cleans data
    - save the data into data group files
    - load the data
    - run many regression algorithms
    
## Base Models

- metric
    - accuracy initially, to check how good it would be to predict correctly

- random guess
    - expected 0.5 accuracy
    
- full with best of first run
    -
