==> Bank/1 - Preprocess.ipynb <==
#Preprocess
This notebook preprocess the data. The basic steps are:

   - Load and join the inputs
   - Split the columns into column groups. That will make it easier to process different sets of columns
   - Preprocess the groups, binarizing, normalizing and creating new features as necessary
   - Save each group into a different file inside **Processed** folder
   - Save the index to the train and test dataset into **Processed** folder
    

import sklearn
import pandas as pd
import mylib.utils as mu
from mylib.utils import (print_time, get_numerical_cols, years_passed, days_passed, 
    save_fields_dict, load_fields_dict, Scaler, binarize)

### Load the input and join databases

# load the datasets
application_raw = pd.read_excel('./input/onsite homework v1.xlsx', 1)
loan_raw = pd.read_excel('./input/onsite homework v1.xlsx', 2)
print_time('Loaded information. Loan #: %d Applications #:%d' %(loan_raw.index.size, application_raw.index.size))

# create the customer id on loan to join with application
print_time('Creating customer_id field to join loan and application')
loan = loan_raw.copy()
loan['customer_id']=loan.apply(lambda l: l[0][:-3].lower(), axis = 1)
loan['loan_no']=loan.apply(lambda l: int(l[0][-2:]), axis = 1)

# found that there is 16 with 2 loans. will keep only the second
idx_more_than_once = (loan.groupby(['customer_id'])['flgGood'].count() >1)
idx_more_than_once = idx_more_than_once[idx_more_than_once == True]
print_time('Number of clients with 2 loans %d' %idx_more_than_once.count())
print_time('Removing first loan. Nr of lines %d' % loan.idLoan.count())
loan.drop(loan[loan.customer_id.isin(idx_more_than_once.index)][loan.loan_no ==1].index, inplace = True)
print_time('Removed first loan. Nr of lines %d' %loan.idLoan.count())

# preprocess application
print_time('Preprocessing application')
application = application_raw.copy()

# create data 
data = pd.merge(application, loan, on='customer_id')
print_time('Final data created. Nr of lines %d' %data.index.size)

# 13 lines lost. Check if it is correct
print_time('Number of lines discarded %d' 
      %application[~application.customer_id.isin(data.customer_id)].customer_id.count())

# target variable
data.rename(columns={'flgGood':'target'}, inplace=True)
data['target'] = [1 if d== 'Good' else 0 for d in data.target]
#del data['target']

### Create group of columns

# cols names helpers
cols=dict(
    raw_scores = [ col for col in data.columns.tolist() if ('raw' in col)],
    personal = [u'Facebook profile duration', u'residence_duration', u'home_phone_type',
                 u'other_phone_type', u'Gender_facebook', u'birth_date', u'Title'],
    location = [u'City', u'State',],
    address = ['CEP', 'StreetAddress', 'Latitude','Longitude'],
    #dates = [u'application_when',  u'birth_date'],
    #ids = ['customer_id', 'idLoan'],
    personal_unique = ['CPF', 'Occupation', 'email', 'TelephoneNumber'],
    name = ['GivenName', 'MiddleInitial', 'Surname'],
    others = ['status'],
    balance = [u'residence_rent_or_own', u'monthly_rent_amount', u'monthly_income_amount', u'bank_account_duration'],
    nu_info = ['loan_no', u'Credit_Line_approved', u'Credit_Line_requested', u'application_when']
)
numeric, non_numeric = get_numerical_cols(data)

# dictionary of transformed data
trans_data = {}
print_time('Feature engineering and cleaning')

# set the index to the id so that all child keep the index
data.set_index('customer_id', inplace=True)

# fill the secondary type with string NI - Not Informed
data.other_phone_type.fillna('NI', inplace=True)
data.bank_account_duration.fillna('NI', inplace=True)
#data.Occupation.fillna('NI', inplace=True)

# binarize categoricals
print_time('Processing personal info')
personal = pd.get_dummies(data[cols['personal']])

# transform date of birth to age
personal['age'] = personal.birth_date.apply(years_passed)
del personal['birth_date']

# create a new feature related to money and age
personal['age_money_ratio_class'] = binarize( data[u'monthly_income_amount']/personal[u'age'])

trans_data['personal'] = personal

# scale the numerical mean 0, sd 1
print_time('Normalizing raw scores')
scl = Scaler(cols['raw_scores'])
raw_scores = scl.fit_transform(data)
trans_data['raw_scores'] = raw_scores[cols['raw_scores']]

print_time('Processing balance information')
balance = data[cols['balance']].copy()
balance.residence_rent_or_own = [1 if r else 0 for r in balance.residence_rent_or_own]
balance['rent_income_ratio'] = balance.monthly_rent_amount/balance.monthly_income_amount
balance = pd.get_dummies(balance)
trans_data['balance'] = balance

print_time('Processing Nu info')
nu_info = data[cols['nu_info']].copy()
nu_info['credit_approved_ratio'] = nu_info.Credit_Line_approved /nu_info.Credit_Line_requested 
nu_info['days_since_application'] = nu_info.application_when.apply(days_passed)
nu_info['credit_income_ratio'] = nu_info.Credit_Line_approved/balance.monthly_income_amount
nu_info['liquid_income_ratio'] = nu_info.Credit_Line_approved/(balance.monthly_income_amount-balance.monthly_rent_amount)
nu_info['money_to_spare_class'] =  binarize(data[u'monthly_income_amount']-data[u'monthly_rent_amount']-nu_info[u'Credit_Line_approved'])
del nu_info['application_when']
trans_data['nu_info'] = nu_info

print_time('Processing location information')
location = data[cols['location']].copy()
location = pd.get_dummies(location)
trans_data['location'] = location

print_time('Creating new features - scores classes')
scores = pd.DataFrame(index=data.index)
scores_list = [u'raw_unit4_score', u'raw_lexisnexis_score',
        u'raw_TU_score', u'raw_FICO_money_score']
names_list = []
for col in scores_list:
    names_list.append(col+'_class')
    scores[col+'_class'] = binarize(data[col])
trans_data['scores_class'] = scores
cols['scores_class'] = names_list

### Save the data to the files

# save full processed
print_time('Saving data in Processed folder')
data.to_csv('./Processed/full_data.csv',  encoding='utf-8')
data['target'].to_csv('./Processed/target.csv',  encoding='utf-8', header=True)

# save files for each set of generated colums
for k,v in cols.iteritems():
    # save transformed data if exists, else save the original one
    if k in trans_data:
        # replace ' ' from column names with _
        trans_data[k].columns = [ col.replace(' ', '_').replace('-', '_').replace('+', '') 
                                 for col in trans_data[k].columns.tolist() ]
        trans_data[k].to_csv('./Processed/{}.csv'.format(k), encoding='utf-8')
    else:
        data[v].to_csv('./Processed/{}.csv'.format(k), encoding='utf-8')

# save the dictionary for future use
save_fields_dict(cols)

### Save the common seed

#56909 seed to run the example
mu.save_seed(56909)
