==> Bank/2 - ML Search.ipynb <==
#ML Search

This notebook is the search for the best ML algorithm for the database and the best combination of features,  differently from the Grid Search from SKLearn lib that search over parameters of one algorithm only.

The basic flow here is:

- create a combinations of all the column groups
- for each combination
    - read the appropriate data 
    - run a series of machine learning algorithms
    - order the algorithms by the best results
    - save the prediction of the 2 best algorithms

%pylab inline
import sklearn
import pandas as pd
import mylib.utils as mu
from itertools import combinations 
import seaborn as sns
from mylib.utils import print_time

groups = ['balance', 'location', 'nu_info', 'personal', 'raw_scores', 'scores_class']
all_combinations = [ p for i in range(len(groups)) for p in combinations(groups, i)][1:]
target = mu.load_target_data()

### Creating a benchmark
Checking what would be the result if the predictions are all 0 or all 1

# split the target to get the validation 
X_train, X_val, y_train, y_val = mu.split_data(target, target)
print_time('All Zeros')
mu.classification_report_matrix(y_val, [0]*len(y_val))
print_time('All Ones')
mu.classification_report_matrix(y_val, [1]*len(y_val))
print_time('Random Guess')
print_time( mu.np.mean([ mu.f1_scorer(y_val, mu.np.random.randint(2, size=len(y_val))) for _ in range(100)]) )

### Search over all combinations

proc_name = 'combination'
best_score = 0
best_run = None
best_models = {}
for comb in all_combinations:
    # load only the data from the combination
    print_time('Processing combination {}'.format(comb))
    data = mu.load_data(comb)
    results = mu.train_regression(data, mu.np.ravel(target), scorer=mu.f1_scorer)
    
    # evaluate the best model agains the test set
    run =  results[0][1]
    run['comb'] = comb
    score_test = run['score_val']
    print_time('Test score {}'.format(score_test))
    
    # save the best model
    if best_score<score_test:
        best_score = score_test
        best_run = run
    
    # save models with score test more than .65
    best_models[score_test] = run

print_time('Best run is:')
print_time('Comb: {}'.format(best_run['comb']))
print_time('Model: {}'.format(best_run['model']))
print_time('Scores: train - {} :: validation - {}'.
           format(best_run['score_train'], best_run['score_val']))
mu.classification_report_matrix(best_run['y_val'], best_run['pred_val'])
mu.save_predictions_from_model(best_run, 'mlsearch_{}'.format(best_run['comb']))

num_best = 3
for score in sorted(best_models.keys())[-num_best-2:-2]:
    model = best_models[score]
    print_time('Comb: {}'.format(model['comb']))
    print_time('Model: {}'.format(model['model']))
    print_time('Scores: train - {} :: validation - {}'.
               format(model['score_train'], model['score_val']))
    mu.classification_report_matrix(model['y_val'], model['pred_val'])
    mu.save_predictions_from_model(model, 'mlsearch_{}'.format(model['comb']))

### Dimensionality redution
The main table has more than 260 columns, for 634 lines. That makes the model overfit. Applying PCA there may be a benefical reduction of the number of columns.

from sklearn.decomposition import PCA, RandomizedPCA 

data = mu.load_data(['balance', 'location', 'nu_info', 'personal', 'raw_scores', 'scores_class'])
pca = RandomizedPCA()

pca_train = pca.fit_transform(data)

results = mu.train_regression(pca_train, ravel(target))
mu.print_best(results, 2)
