==> Bank/2.2 - Knn.ipynb <==
%pylab inline
import sklearn
import pandas as pd
import mylib.utils as mu
from sklearn.neighbors import KNeighborsClassifier

proc_name = 'knn'
data = mu.load_data(['balance', 'location', 'nu_info', 'personal', 'raw_scores'])

target = mu.load_target_data()

for w in ['uniform', 'distance', lambda x: np.log(x)]:
    models = [KNeighborsClassifier(n_neighbors=11, weights=w)]
    results = mu.train_regression(data, ravel(target), seed=333, models=models)
    mu.print_best(results)

# save the original y and the predicted
mu.save_predictions_from_all(results, run_name=proc_name)

y_val = pd.DataFrame(dict(pred=results[1][1]['y_val']))
y_val.to_csv('./Output/y_val.csv', index=False)

from sklearn.metrics import confusion_matrix, classification_report
print classification_report(results[1][1]['y_val'], results[0][1]['pred_val'])
print confusion_matrix(results[1][1]['y_val'], results[0][1]['pred_val'])

import seaborn as sns
model = results[0][1]['model']
# print model info
f, ax = plt.subplots(figsize=(15, 15))
index_most_important = model.feature_importances_>0.03
ax = sns.barplot(x=data.columns[index_most_important], y=model.feature_importances_[index_most_important])
