==> Bank/4.1 - ANN - Lasagne.ipynb <==
import mylib.utils as mu
from lasagne.layers import InputLayer, DenseLayer
from lasagne.nonlinearities import softmax, rectify
from lasagne.objectives import binary_crossentropy
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import StandardScaler
from lasagne.layers import DenseLayer
from lasagne.layers import InputLayer
from lasagne.layers import DropoutLayer
from lasagne.nonlinearities import softmax, sigmoid, tanh
from lasagne.updates import nesterov_momentum
from nolearn.lasagne import NeuralNet

import theano
import numpy as np
# load data
comb = ['balance', 'scores_class']
data = mu.load_data(comb)
target = mu.load_target_data()

# split the data to check the training
X_train, X_val, y_train, y_val = mu.split_data(data, target)

X = X_train.values.copy()
y = y_train.astype(np.int32)
num_classes = 2
num_features = X.shape[1]
print(num_classes, num_features)

layers0 = [('input', InputLayer),
           ('dense0', DenseLayer),
           #('dense1', DenseLayer),
           ('output', DenseLayer)]

net0 = NeuralNet(layers=layers0,
                 
                 input_shape=(None, num_features),
                 dense0_num_units=400,
                 #dropout_p=0.5,
                 #dense1_num_units=200,
                 output_num_units=num_classes,
                 output_nonlinearity=softmax,
                 #objective_loss_function= binary_crossentropy,
                 
                 update=nesterov_momentum,
                 update_learning_rate=0.01,
                 update_momentum=0.9,
                 
                 eval_size=0.2,
                 verbose=1,
                 max_epochs=40)

net0.fit(X.astype(np.float32), np.ravel(y))

y_prob = net0.predict_proba(X_val.values.astype(np.float32))

y_prob
