==> Bank/5 - Bagging.ipynb <==
# Bagging the ouputs
This script will get all the outputs that were generated from other processes and create a new one. The output for each sample is based on majority voting of the outputs. This is based on the idea of [Kaggle Ensembling Guide](http://mlwave.com/kaggle-ensembling-guide/).

For the bagging, a function was created that will read all the files using glob string and create the final preditiction output. There is an option to read only the files that had higher accuracy than a value

%pylab inline
import sklearn
import pandas as pd
import mylib.utils as mu

### Create the output file based on voting

# load any data, just to get 
data = mu.load_data(['scores_class'], del_key=False)
target = mu.load_target_data()
X_train, X_val, y_train, y_val = mu.split_data(data, target)
mu.bag_results('./Output/pred_*.csv', './Output/bagged_pred.csv', score_filter=None)

# read both files and do the calculus
pred = pd.read_csv('./Output/bagged_pred.csv')
# print the results
mu.classification_report_matrix(y_val, pred.preds)
mu.print_time(mu.f1_scorer(y_val.target, pred.preds))

### No VowpalWabbit

# load any data, just to get 
data = mu.load_data(['scores_class'], del_key=False)
target = mu.load_target_data()
X_train, X_val, y_train, y_val = mu.split_data(data, target)
mu.bag_results('./Output/pred_mlsearch*.csv', './Output/bagged_pred.csv', score_filter=None)

# read both files and do the calculus
pred = pd.read_csv('./Output/bagged_pred.csv')
# print the results
mu.classification_report_matrix(y_val, pred.preds)
mu.print_time(mu.f1_scorer(y_val.target, pred.preds))
